+++ AdventOfCode2023/Utils/PathFinding/BreadthFirst.cs
@@ -5,7 +5,7 @@ namespace AdventOfCode2023.Utils.Pathfinding
     public class BreadthFirst<TNode> : IPathFinder<TNode> where TNode : IEquatable<TNode>, IComparable<TNode>
     {
         public bool HasSolution { get; }
-        public int TotalCost { get; }
+        public long TotalCost { get; }
         public List<TNode> Path { get; }
         public Dictionary<TNode, int> DistancesMap { get; }
