+++ AdventOfCode2023/Utils/PathFinding/IPathFinder.cs
@@ -5,7 +5,7 @@ namespace AdventOfCode2023.Utils.Pathfinding
     public interface IPathFinder<TNode> where TNode : IEquatable<TNode>, IComparable<TNode>
     {
         public bool HasSolution { get; }
-        public int TotalCost { get; }
+        public long TotalCost { get; }
         public List<TNode> Path { get; }
     }
 }
